% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootpval.R
\name{bootpval}
\alias{bootpval}
\title{Bootstrap of P value}
\usage{
bootpval(x, conf.level = 0.95, iter = 3000, mu0 = 0, test = "two")
}
\arguments{
\item{x}{data}

\item{conf.level}{confidence level}

\item{iter}{number of iterations}

\item{mu0}{null value}

\item{test}{type of test}
}
\description{
Bootstrap of P value
}
