% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set}
\format{
A data frame with 144 rows and 5 columns
\describe{
  \item{LENGTH}{Length of fish}
  \item{WEIGHT}{Weight of fish}
  \item{SPECIES}{Species of Fish}
  \item{DDT}{amount of ddt in fish}\
  \item{RIVER}{which river fish was found}
  \item{MILE}{How far in river where fish was caught}
}
}
\source{
<https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850>
}
\usage{
ddt
}
\description{
A complete dat set ofmeasuremetns of 144 fish caught in various rivers
Report ...
}
\keyword{datasets}
