% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage}
\format{
A data frame with 15 rows and 2 variables
\describe{
  \item{DISTANCE}{a distance from fire station in miles}
  \item{DAMAGE}{X 1000 dollars in dammage}
}
}
\source{
<https://www.crcpress.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850>
}
\usage{
fire
}
\description{
Data set collected concerning damage from fire
Report ...
}
\keyword{datasets}
