% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Binomial Histogram}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterations}

\item{n}{sample size}

\item{p}{probability}
}
\value{
A histogram
}
\description{
Binomial Histogram
}
\examples{
\dontrun{mybin(iter = 100, n = 10, p = 0.5))}
}
