% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Procedure for statistics}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample}

\item{fun}{statistic of interest}

\item{alpha}{probability of tail}

\item{cx}{cex point size}
}
\description{
Bootstrap Procedure for statistics
}
\examples{
\dontrun{myboot(iter=10000, x=sam, fun="mean", alpha = 0.05, cx =1.5))}
}
