% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval Function}
\usage{
myci(x, alpha)
}
\arguments{
\item{x}{sample}

\item{alpha}{type 1 error}
}
\description{
Confidence Interval Function
}
\examples{
 \dontrun{myci(x=sam, alpha = 0.05)}
}
