% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltb.R
\name{mycltb}
\alias{mycltb}
\title{Normal Approximation of Binomial Distribution}
\usage{
mycltb(n, iter, p = 0.5)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{p}{probability of success}
}
\description{
Normal Approximation of Binomial Distribution
}
\examples{
\dontrun{mycltb(n=100, iter = 1000, p = 0.5)}
}
