% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Max Likelihood Function}
\usage{
mymaxlik(lfun, x, param)
}
\arguments{
\item{lfun}{log function}

\item{x}{sample of ys}

\item{param}{seq of probabilities}
}
\description{
Max Likelihood Function
}
\examples{
\dontrun{mymaxlik(x=c(3,4,5,2,3,4,5,5,5,1),param=seq(0,1,length=1000),lfun=logbin)}
}
