% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Lower Tail Value}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{value in normal distribution}

\item{mu}{mean}

\item{sigma}{standard deviation}
}
\value{
Shaded Region of Lower Tail Value
}
\description{
Normal Lower Tail Value
}
\examples{
\dontrun{myncurve(0.5, 0, 1)}
}
