% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Overbooking Ticket Function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{mean}

\item{gamma}{probability plane will be overbooked}

\item{p}{probability of a show}
}
\value{
Binomial solution with nd as binomial answer to number of seats to sell with a plot along with a Normal approximation answer
}
\description{
Overbooking Ticket Function
}
\examples{
\dontrun{ntickets(200, 0.02, 0.95)}
}
