% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterhist.R
\name{scatterhist}
\alias{scatterhist}
\title{Scatter histogram}
\usage{
scatterhist(x, y, xlab = "", ylab = "")
}
\arguments{
\item{x}{x variable as vector}

\item{y}{y variable as vector}

\item{xlab}{x label}

\item{ylab}{y label}
}
\value{
A sophisticated plot
}
\description{
Scatter histogram
}
\examples{
\dontrun{scatterhist(ddt$LENGTH, ddt$WEIGHT, xlab = "Length")}
}
